# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin API 文档",
        "title": "Gin Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "25076778@qq.com"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/api/v1/menu": {
            "get": {
                "description": "菜单列表",
                "tags": [
                    "菜单管理"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/request.PageData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.Menu"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建菜单",
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/request.MenuCreate"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/{id}": {
            "get": {
                "description": "菜单详情",
                "tags": [
                    "菜单管理"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "更新菜单",
                "tags": [
                    "菜单管理"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.MenuUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "菜单删除",
                "tags": [
                    "菜单管理"
                ],
                "summary": "菜单删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/role/{id}/menu": {
            "get": {
                "description": "角色菜单",
                "tags": [
                    "菜单管理"
                ],
                "summary": "角色菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "角色ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.ArgsErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Invalid arguments"
                }
            }
        },
        "errcode.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "errcode.SystemErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "model.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子节点",
                    "type": "array",
                    "items": {}
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "menuAction": {
                    "description": "菜单功能",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MenuActions"
                        }
                    ]
                },
                "meta": {
                    "description": "菜单元数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MenuMeta"
                        }
                    ]
                },
                "name": {
                    "description": "路由名称",
                    "type": "string"
                },
                "pid": {
                    "description": "父级id",
                    "type": "integer"
                },
                "roleMenus": {
                    "description": "角色菜单",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoleMenus"
                    }
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态 1=启用 2=停用",
                    "type": "integer"
                },
                "type": {
                    "description": "类型 1=菜单 2=功能",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "model.MenuActions": {
            "type": "object",
            "properties": {
                "authValue": {
                    "type": "string"
                },
                "btnSize": {
                    "type": "string"
                },
                "btnStyle": {
                    "type": "string"
                },
                "btnType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isConfirm": {
                    "type": "integer"
                },
                "isLink": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "transKey": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.MenuMeta": {
            "type": "object",
            "properties": {
                "authBtnList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MenuActions"
                    }
                },
                "component": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isAffix": {
                    "type": "integer"
                },
                "isHide": {
                    "type": "integer"
                },
                "isIframe": {
                    "type": "integer"
                },
                "isKeepAlive": {
                    "type": "integer"
                },
                "isLink": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoleMenus"
                    }
                },
                "title": {
                    "type": "string"
                },
                "transKey": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.RoleMenus": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menuId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roleId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "fullName": {
                    "description": "姓名",
                    "type": "string"
                },
                "gender": {
                    "description": "性别 1=男 2=女",
                    "type": "integer"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "mainDept": {
                    "$ref": "#/definitions/model.UserDepartments"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "status": {
                    "description": "状态 1=启用 2=停用",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userDepts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserDepartments"
                    }
                },
                "userRoles": {
                    "description": "用户角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserRoles"
                    }
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "model.UserDepartments": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "integer"
                },
                "dept": {
                    "$ref": "#/definitions/model.Department"
                },
                "id": {
                    "type": "integer"
                },
                "isMain": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "model.UserRoles": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "roleId": {
                    "description": "角色id",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "userId": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
        "request.ActionCreate": {
            "type": "object",
            "properties": {
                "authValue": {
                    "type": "string"
                },
                "btnSize": {
                    "type": "string"
                },
                "btnStyle": {
                    "type": "string"
                },
                "btnType": {
                    "type": "string"
                },
                "isConfirm": {
                    "type": "integer"
                },
                "isLink": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "transKey": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "request.MenuCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "isLink": {
                    "type": "integer"
                },
                "menuAction": {
                    "$ref": "#/definitions/request.ActionCreate"
                },
                "meta": {
                    "$ref": "#/definitions/request.Meta"
                },
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "roleMenus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.RoleMenu"
                    }
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "request.MenuUpdate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "menuAction": {
                    "$ref": "#/definitions/request.ActionCreate"
                },
                "meta": {
                    "$ref": "#/definitions/request.Meta"
                },
                "name": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "roleMenus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.RoleMenu"
                    }
                },
                "sort": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "request.Meta": {
            "type": "object",
            "required": [
                "component",
                "icon",
                "path",
                "title"
            ],
            "properties": {
                "component": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "isAffix": {
                    "type": "integer"
                },
                "isHide": {
                    "type": "integer"
                },
                "isIframe": {
                    "type": "integer"
                },
                "isKeepAlive": {
                    "type": "integer"
                },
                "isLink": {
                    "type": "string"
                },
                "menuId": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "redirect": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "transKey": {
                    "type": "string"
                }
            }
        },
...
        "request.PageData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "数据列表"
                },
                "notPage": {
                    "description": "不分页",
                    "type": "boolean",
                    "example": false
                },
                "page": {
                    "description": "当前页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "request.RoleMenu": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "menuId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "roleId": {
                    "type": "integer"
                }
            }
        },
...
    }
}